Envrironment update
Minor changes to the collision system with the introduction of new lists
for different types of objects such as envionement and player.

The platform code has now been added to the game and is now awaiting the
different typesof platforms we need for the game levels.

Player can now jump onto the crate now and whenever the player or the
crates leave a platform, their gravity bool is set to true